introduce copy on write lists instead of read only lists
